+++ src/packages/inputnumber/demos/h5/demo1.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { InputNumber } from '@nutui/nutui-react'
+
+const Demo1 = () => {
+  return <InputNumber defaultValue={1} allowEmpty />
+}
+export default Demo1
