+++ src/packages/inputnumber/demos/h5/demo4.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { InputNumber } from '@nutui/nutui-react'
+
+const Demo4 = () => {
+  return <InputNumber defaultValue={0} disabled />
+}
+export default Demo4
