+++ src/packages/inputnumber/demos/h5/demo5.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { InputNumber } from '@nutui/nutui-react'
+
+const Demo5 = () => {
+  return <InputNumber defaultValue={1} readOnly />
+}
+export default Demo5
