+++ src/packages/inputnumber/demos/taro/demo2.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { InputNumber } from '@nutui/nutui-react-taro'
+
+const Demo2 = () => {
+  return <InputNumber defaultValue={0} min={0} step={5} />
+}
+export default Demo2
