+++ src/packages/inputnumber/demos/taro/demo4.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { InputNumber } from '@nutui/nutui-react-taro'
+
+const Demo4 = () => {
+  return <InputNumber defaultValue={0} disabled />
+}
+export default Demo4
