+++ src/packages/inputnumber/demos/taro/demo5.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { InputNumber } from '@nutui/nutui-react-taro'
+
+const Demo5 = () => {
+  return <InputNumber defaultValue={1} readOnly />
+}
+export default Demo5
