+++ src/packages/inputnumber/demos/taro/demo7.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { InputNumber } from '@nutui/nutui-react-taro'
+
+const Demo7 = () => {
+  return <InputNumber defaultValue={5.5} step={0.1} digits={1} readOnly />
+}
+export default Demo7
